Fix bug with refresh main page on create transaction
